+++ src/server/utils.ts
@@ -1,9 +1,10 @@
-import fs from 'fs';
+import fs, { createWriteStream, existsSync, readFileSync, unlink } from 'fs';
 import path from 'path';
 import cluster from 'cluster';
-import { SkipOptions, isDefined, Test } from '../types';
+import { SkipOptions, isDefined, Test, noop } from '../types';
 import { emitShutdownMessage, sendShutdownMessage } from './messages';
 import findCacheDir from 'find-cache-dir';
+import { get } from 'https';
 
 export const LOCALHOST_REGEXP = /(localhost|127\.0\.0\.1)/i;
 
@@ -91,17 +92,25 @@ export function requireConfig<T>(configPath: string): T {
   return configModule && configModule.__esModule ? configModule.default : configModule;
 }
 
-export function shutdownWorkers(): void {
+export async function shutdownWorkers(): Promise<void> {
   emitShutdownMessage();
-  Object.values(cluster.workers)
-    .filter(isDefined)
-    .filter((worker) => worker.isConnected())
-    .forEach((worker) => {
-      const timeout = setTimeout(() => worker.kill(), 10000);
-      worker.on('exit', () => clearTimeout(timeout));
-      sendShutdownMessage(worker);
-      worker.disconnect();
-    });
+  await Promise.all(
+    Object.values(cluster.workers)
+      .filter(isDefined)
+      .filter((worker) => worker.isConnected())
+      .map(
+        (worker) =>
+          new Promise<void>((resolve) => {
+            const timeout = setTimeout(() => worker.kill(), 10000);
+            worker.on('exit', () => {
+              clearTimeout(timeout);
+              resolve();
+            });
+            sendShutdownMessage(worker);
+            worker.disconnect();
+          }),
+      ),
+  );
 }
 
 export function getStorybookVersion(): string {
@@ -121,7 +130,7 @@ export function isStorybookVersionLessThan(major: number, minor?: number): boole
 }
 
 export function getCreeveyCache(): string {
-  return findCacheDir({ name: 'creevey' }) as string;
+  return findCacheDir({ name: 'creevey', cwd: __dirname }) as string;
 }
 
 export async function runSequence(seq: Array<() => unknown>, predicate: () => boolean): Promise<void> {
@@ -143,3 +152,33 @@ export function testsToImages(tests: (Test | undefined)[]): Set<string> {
     ),
   );
 }
+
+// https://tuhrig.de/how-to-know-you-are-inside-a-docker-container/
+export const isInsideDocker = existsSync('/proc/1/cgroup') && /docker/.test(readFileSync('/proc/1/cgroup', 'utf8'));
+
+export const downloadBinary = (downloadUrl: string, destination: string): Promise<void> =>
+  new Promise((resolve, reject) =>
+    get(downloadUrl, (response) => {
+      if (response.statusCode == 302) {
+        const { location } = response.headers;
+        if (!location)
+          return reject(new Error(`Couldn't download selenoid. Status code: ${response.statusCode ?? 'UNKNOWN'}`));
+
+        return resolve(downloadBinary(location, destination));
+      }
+      if (response.statusCode != 200)
+        return reject(new Error(`Couldn't download selenoid. Status code: ${response.statusCode ?? 'UNKNOWN'}`));
+
+      const fileStream = createWriteStream(destination);
+      response.pipe(fileStream);
+
+      fileStream.on('finish', () => {
+        fileStream.close();
+        resolve();
+      });
+      fileStream.on('error', (error) => {
+        unlink(destination, noop);
+        reject(error);
+      });
+    }),
+  );